move Trunc, Cot, Round, iszero functions to datafusion-functions
