+++ datafusion/proto/proto/datafusion.proto
@@ -555,12 +555,12 @@ enum ScalarFunction {
   Log = 11;
   // 12 was Log10
   // 13 was Log2
-  Round = 14;
+  // 14 was Round
   // 15 was Signum
   // 16 was Sin
   // 17 was Sqrt
   // Tan = 18;
-  Trunc = 19;
+  // 19 was Trunc
   // 20 was Array
   // RegexpMatch = 21;
   // 22 was BitLength
@@ -642,7 +642,7 @@ enum ScalarFunction {
   // 98 was Cardinality
   // 99 was ArrayElement
   // 100 was ArraySlice
-  Cot = 103;
+  // 103 was Cot
   // 104 was ArrayHas
   // 105 was ArrayHasAny
   // 106 was ArrayHasAll
@@ -653,7 +653,7 @@ enum ScalarFunction {
   Nanvl = 111;
   // 112 was Flatten
   // 113 was IsNan
-  Iszero = 114;
+  // 114 was Iszero
   // 115 was ArrayEmpty
   // 116 was ArrayPopBack
   // 117 was StringToArray
